Add || true to .bashrc lines in case set -e is on
In the (super-rare) case that someone does something like:

```bash
set -e
. ~/.bashrc
```

and the nvm lines are the last two lines in the shell script, the bashrc
will return 1 as the exit code, killing the parent shell.

An alternative workaround for this is to add `; true` (or `; :` for
maximum opacity) instead of `|| true`), but I feel like `|| true` is
most obvious.